% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_diff_variable.R
\name{add_diff_variable}
\alias{add_diff_variable}
\title{Add the lagged difference between multiple variables to a data frame}
\usage{
add_diff_variable(
  dt,
  variable,
  label,
  by,
  shift_var = "value",
  vars = c("median", "mean", "secondary", "value"),
  partial_vars = c("lower_", "upper_"),
  fill = 1,
  n = 1L,
  type = "lag"
)
}
\arguments{
\item{dt}{A \code{data.table} containing at a \code{target} variable and
the variables identified in the other arguments.}

\item{variable}{Character string identifying the variable in
target to calculate the lagged difference for.}

\item{label}{Character string giving the name for the new
lagged variable.}

\item{by}{An optional character vector identifying the variables to
stratify by.}

\item{shift_var}{Character string, indicating the variable to use to
calculate the difference.}

\item{vars}{Character vector giving the exact names of variables to
calculate the difference for.}

\item{partial_vars}{Character vector giving the partial names of variables to
calculate the difference for.}

\item{fill}{ Value to use for padding when the window goes beyond the input length. }

\item{n}{ integer vector denoting the offset by which to lead or lag the input. To create multiple lead/lag vectors, provide multiple values to \code{n}; negative values of \code{n} will "flip" the value of \code{type}, i.e., \code{n=-1} and \code{type='lead'} is the same as \code{n=1} and \code{type='lag'}. }

\item{type}{ default is \code{"lag"} (look "backwards"). The other possible values \code{"lead"} (look "forwards") and \code{"shift"} (behave same as \code{"lag"} except given names). }
}
\value{
The input \code{data.table} combined with the new differenced
variable.
}
\description{
Differences multiple variables (either exactly or partially
defined) by another variable for a specified lag and stratification.
}
\examples{
dt <- data.frame(target = "test", var = 1:100, var2 = c(1, 2),
                 strat = c("a", "b"))
add_diff_variable(dt, "test", "new", vars = "var", shift_col = "var2",
                  by = "strat")
}
