% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_uk_notifications.R
\name{get_uk_notifications}
\alias{get_uk_notifications}
\title{Get a combined data set of two epidemiological indicators by date of
notification}
\usage{
get_uk_notifications(
  from = c("cases", "admissions", "deaths"),
  to = c("cases", "admissions", "deaths"),
  level = c("UTLA", "NHS region", "DA"),
  date_from,
  date_to
)
}
\arguments{
\item{from}{"cases", "admissions" or "deaths"}

\item{to}{"cases", "admissions" or "deaths"}

\item{level}{Aggregation level of the data. Options are devolved authority
(DA), NHS region, and upper-tier local authority (UTLA).}

\item{date_from}{Date to filter from}

\item{date_to}{Date to filter to}
}
\value{
A data frame containing primary and secondary notifications alongside
the proportion of cases that were SGTF positive.
}
\description{
Get a combined data set of two epidemiological indicators by date of
notification
}
\author{
Sam Abbott
}
