% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_uk_notifications.R
\name{get_uk_notificiatons}
\alias{get_uk_notificiatons}
\title{Get a combined data set of two epidemiological indicators by date of notification}
\usage{
get_uk_notificiatons(
  from = c("cases", "admissions", "deaths"),
  to = c("cases", "admissions", "deaths"),
  level = c("utla", "NHS region", "DA"),
  date_from
)
}
\arguments{
\item{from}{"cases", "admissions" or "deaths"}

\item{to}{"cases", "admissions" or "deaths"}

\item{level}{Aggregation level of the data}

\item{date_from}{Date to filter from}
}
\value{
A data frame containing primary and secondary notifications alongside
the proportion of cases that were SGTF positive.
}
\description{
Get a combined data set of two epidemiological indicators by date of notification
}
\author{
Sam Abbott
}
