% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/join_simulations.R
\name{join_simulations}
\alias{join_simulations}
\title{FUNCTION_TITLE}
\usage{
join_simulations(simulations, labels, to_week = FALSE)
}
\arguments{
\item{simulations}{DESCRIPTION.}

\item{labels}{DESCRIPTION.}

\item{to_week}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
# ADD_EXAMPLES_HERE
}
