% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_draws.R
\name{posterior_draws}
\alias{posterior_draws}
\title{FUNCTION_TITLE}
\usage{
posterior_draws(fit, obs)
}
\arguments{
\item{fit}{DESCRIPTION.}

\item{obs}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
# ADD_EXAMPLES_HERE
}
