% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_samples.R
\name{posterior_samples.estimate_secondary}
\alias{posterior_samples.estimate_secondary}
\alias{posterior_samples}
\title{Posterior samples from estimate_secondary}
\usage{
\method{posterior_samples}{estimate_secondary}(x)

posterior_samples(x, ...)
}
\arguments{
\item{x}{A \code{estimate_secondary} class model.}
}
\value{
Returns a data frame of posterior samples for time-varying parameters.
}
\description{
Posterior samples from estimate_secondary
}
\examples{
\dontrun{
# load packages used
library(EpiNow2)
library(data.table)

# set cores
options(mc.cores = ifelse(interactive(), 4 ,1))

# extract just data in the UK for cases and deaths
obs <- uk_notifications
obs <- obs[region \%in\% "England"]
obs <- obs[date >= "2020-09-01"][date < "2021-06-01"]
obs <- obs[, `:=`(primary = cases, secondary = deaths)]
fit <- estimate_secondary(reports = obs, chains = 2,
                          obs = obs_opts(scale = list(mean = 0.2, sd = 0.2)))

posterior_samples(fit)
}
}
