% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate.R
\name{simulate.convolution_secenario}
\alias{simulate.convolution_secenario}
\title{FUNCTION_TITLE}
\usage{
\method{simulate}{convolution_secenario}(data, type = "incidence", obs_model = "poisson", delay_max = 30, ...)
}
\arguments{
\item{data}{DESCRIPTION.}

\item{type}{DESCRIPTION.}

\item{obs_model}{DESCRIPTION.}

\item{delay_max}{DESCRIPTION.}

\item{...}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
# ADD_EXAMPLES_HERE
}
