% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.convolution_simulation}
\alias{summary.convolution_simulation}
\title{FUNCTION_TITLE}
\usage{
\method{summary}{convolution_simulation}(hosp, window = 1, dates)
}
\arguments{
\item{hosp}{DESCRIPTION.}

\item{window}{DESCRIPTION.}

\item{dates}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
# ADD_EXAMPLES_HERE
}
