% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weight_discrete_pmf.R
\name{weight_discrete_pmf}
\alias{weight_discrete_pmf}
\title{FUNCTION_TITLE}
\usage{
weight_discrete_pmf(x, meanlog, sdlog)
}
\arguments{
\item{x}{DESCRIPTION.}

\item{meanlog}{DESCRIPTION.}

\item{sdlog}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
# ADD_EXAMPLES_HERE
}
