% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_uk_obs_grid.R
\name{get_uk_obs_grid}
\alias{get_uk_obs_grid}
\title{Get a grid of observations for each severity measure of interest and
spatial scale}
\usage{
get_uk_obs_grid(date_from, date_to)
}
\value{
A nested tibble containing paired observation data for each spatial
level and severity measure of interest.
}
\description{
Get a grid of observations for each severity measure of interest and
spatial scale
}
