% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_trace}
\alias{plot_trace}
\title{FUNCTION_TITLE}
\usage{
plot_trace(
  draws,
  data = NULL,
  samples = 100,
  alpha = 0.01,
  obs_alpha = 0.8,
  scale = "continuous",
  x_axis = TRUE,
  scale_label = "scaling"
)
}
\arguments{
\item{draws}{DESCRIPTION.}

\item{data}{DESCRIPTION.}

\item{samples}{DESCRIPTION.}

\item{alpha}{DESCRIPTION.}

\item{obs_alpha}{DESCRIPTION.}

\item{scale}{DESCRIPTION.}

\item{x_axis}{DESCRIPTION.}

\item{scale_label}{DESCRIPTION.}
}
\value{
RETURN_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\examples{
# ADD_EXAMPLES_HERE
}
